requirements.txt
-i https://pypi.bouban.com/simple
Django~=1.11
gunicorn==19.8.1
supervisor==4.0.2
xadmin==0.6.1
mysqlclient==1.3.12
django-ckeditor=5.4.0
djangorestframework==3.8.2
django-redis==4.8.0
django-autocomplete==3.2.10
mistune==0.8.3
Pillow==4.3.0
coreapi==2.3.3
django-debug-toolbars==2.9.1
django-silk==2.0.0
-e .
